tests: media recorder backend - silence -Wsign-compare
Making (apple)clang happy
Amends a14c64dcec54cd9846a5ec3db5babf2d6c1b932c

Pick-to: 6.12

+++ tests/auto/integration/qmediarecorderbackend/tst_qmediarecorderbackend.cpp
@@ -76,7 +76,7 @@ std::set<QMediaFormat::VideoCodec> unsupportedVideoCodecs(QMediaFormat::FileForm
 
 void verifyEncodedFrames(const MediaInfo &info, const std::vector<QColor> &expectedColors)
 {
-    QCOMPARE_EQ(info.m_frameCount, expectedColors.size());
+    QCOMPARE_EQ(size_t(info.m_frameCount), expectedColors.size());
     QCOMPARE_EQ(info.m_colors.size(), expectedColors.size());
     for (auto i = 0U; i < expectedColors.size(); ++i)
         QVERIFY2(fuzzyCompare(info.m_colors[i][0], expectedColors.at(i)),
